test: update test-run (unix socket path length)
See the previous commit 'test: descrease instance file name length' for
the problem description.

Since we're going to use the tags like 2.10.0-beta1 for upcoming
pre-releases, we have even less symbols for a filename of a unix socket.
So I chose the shorter names on the test-run side: foo.socket-iproto ->
foo.i, foo.socket-admin -> foo.c (console).

See https://github.com/tarantool/test-run/pull/312

This update is mostly necessary for the current master branch, but it'll
be applied for 1.10 as well: we keep test-run consistent across alive
branches.

Part of #6184