github-ci:Create new Telegram Notifier action The problem was that previous action worked 50/50 or did not work at all. So I decided to create a new action which is functional and has the same process logic
Also two additional actions were added. They are required
for telegram action work better. First one is setting
additional environment variables that we are need for
message to be correctly structured and another is to check
if commit sha exist or not

All explanations can be found in README.md of each action