sql: fix comparison between DECIMAL and big DOUBLE
This patch fixes the comparison between DECIMAL as a left value and
DOUBLE greater than or equal to 1e38 or less than or equal to -1e38 as a
right value. Any DOUBLE value greater than or equal to 1e38 is now
greater than any DECIMAL value, and a DOUBLE value less than or equal to
-1e38 is less than any DECIMAL value. This is because our decimal cannot
contain more than 38 digits.

Closes #6376