sql: fix error on copy empty string in mem_copy()
This patch fixes the problem with copying an empty string in mem_copy().
Previously, because the string length was 0, an error was thrown, but
the diag was not set, which could lead to an error due to an empty diag
or to a double free.

Closes #6157
Closes #6399

(cherry-picked from 2f21f3232273a530851a66bb8ed638213ad9eaf8)