access: fix diag raised when privileges are granted not from admin
There are two kind of schema objects: objects themselves (featuring
specific ids) and entities (representing all kinds of objects of the
concrete type). Granting privileges for entities from user different from
admin is not allowed. In this case diag is raised. Accidentally, there
was a bug in diag: instead of printing entity name attempt to print
object name took place. That led to assertion fault on debug version.
Let's fix this and print entity name as it supposed to be.

Closes #5389