ci: cancel outdated workflow runs
According to a huge amount of commit pushes by developers we should
cancel all outdated workflow runs (previously scheduled and not relevant
due to new changes) to make CI more efficient. GitHub Actions provides
the 'concurrency' feature [1] as a method of reaching that and this
patch introduces its using.

How does it work?

Basically, an update of a developer branch cancels the previously
scheduled workflow run for this branch. However, the 'master' branch,
release branch (1.10, 2.8, etc.), and tag workflow runs are never
canceled.

[1] https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#concurrency

Closes tarantool/tarantool-qa#100