ci: add workflow for checking module integration
For now, there is no testing for the tarantool project verifying its
integration with different modules and connectors from the ecosystem.
This is a quite huge gap in our CI system that is going to be covered
by these changes.

This patch introduces a couple of new workflow files that are named
'integration.yml' and 'reusable_build.yml'. The main workflow is
'integration.yml' that will run automatically per push to master and
release branches. Also, this workflow can be run manually against a
development branch. The 'reusable_build.yml' workflow is called by the
main workflow and builds needed tarantool packages (at this moment for
Ubuntu Focal Fossa only), then stores them as a build artifact. After
that, the main workflow calls the 'reusable_testing.yml' workflow from
a module project that tarantool is going to verify integration with.
The testing workflow should download the tarantool build artifact and
run tests against it.

The basic scheme describing the verification process is represented
below:

      integration.yml
             |
             |
             v
     reusable_build.yml --> <artifact>
             |                  |
             |                  |
             v                  |
    reusable_testing.yml <------+
             |
             |
             v
          <result>

For now, we start only with the vshard module. In the future, we are
going to extend the module list incrementally.

Part of #5265
Part of #6056
Closes #4972

(cherry picked from commit 8ac76fd62115d1c8fb75e6622dd3777f3308a494)