sql: fixed potential OoB issue in lemon.c file
The static analyzer found a potential OoB problem in the lemon.c file.
This situation occurs when you go beyond the boundaries of the
array x in the Rule_sort() function. The number of elements
of the array x used for sorting has a logarithmic dependence on
the size of the input list. Thus, to go beyond the bounds of the array x
and refer to index 32, the size of the input list must be 2^33.

This patch adds an array index check to the Rule_sort () function.
If the 32-item limit is exceeded, the function aborts and
returns NULL instead of the sorted list. The function interrupts its
execution and does not complete the sorting, because the number
of elements in the list exceeding INT32_MAX violates the logic of the program.

It is technically impossible to reproduce this problem in a test, since, firstly,
such a test will require memory allocation for 2^33 elements of the rule structure,
and secondly, it will take too much time to process such a test.

Fixes #5789