iproto: fix kharon use after free
The session can be closed while kharon is travelling between tx and
iproto. If this happens, kharon shouldn't go back to iproto when it
returns to tx, even if there are pending pushes, because the connection
is going to be freed soon.

Thanks to @Gerold103 for the simple fix.

Closes #6520

(cherry picked from commit 2ad640782b2e81872cf4ae4abd4720cb97bb3c0c)