test: fix flaky box-tap/gh-5602-environment-vars-cfg test
The test is flaky, because it tries to listen on a fixed TCP port, which
may be busy. Let's use Unix sockets to fix the test.

While we are at it, let's also remove useless creation of a temporary
directory - the test copies a script there, but never uses it.