net.box: drop establish_connection and wrap methods
establish_connection is used by console to check greeting: if the
protocol is 'Binary', it wraps the connected socket in a net.box
connection; if it's 'Lua console', it wraps the socket in a console
handler.

Presence of these methods complicates migration of the net.box state
machine to C. Also, it'll be difficult to support establish_connection
when we switch net.box to the iostream abstraction, because a greeting
should be received via iostream, not fd, probably after some
protocol-specific negoitiation (e.g. TLS), while iostream isn't (and
won't be) available in Lua and so can't be easily passed around from
establish_connection to Lua and then back to net.box state machine.

Since Lua console will always use plain sockets (it doesn't require
encryption as it's supposed to run on the local host), let's first try
to establish a net.box connection and if it fails because of unsupported
protocol, try to establish a Lua console connection. It means that the
socket will be connected twice for Lua console, which is suboptimal, but
it should be fine, because it's not a hot path.