test: fix flaky wal_off/snapshot_stress test
This commit fixes the following failure:

```
[028] --- wal_off/snapshot_stress.result	Fri Jun  5 06:49:36 2020
[028] +++ wal_off/snapshot_stress.reject	Fri Oct 16 13:48:23 2020
[028] @@ -375,7 +375,7 @@
[028]  ...
[028]  snaps_find_status;
[028]  ---
[028] -- snaps found
[028] +- where are my snapshots?
[028]  ...
[028]  snapshot_check_failed = false
[028]  while #snaps > initial_snap_count do
[028]
```

It happens, because the fiber making snapshots is started after worker
fibers and it works until all worker fibers are done so it might occur
that it doesn't make any snapshots. Fix this by replacing the while-loop
with a repeat-loop, thus ensuring that the worker fiber makes at least
one snapshot.

Closes #5431