error: report ER_READONLY reason in message
A previous commit added reason details as error fields. But error
objects are serialized as strings by default thus loosing all the
details.

This commit makes ER_READONLY message contain the reason why it
happened in a human-readable form. That way it will be well
visible in the logs at least.

Follow-up #5568