net.box: raise if on_event callback failed
The callback isn't supposed to normally raise exceptions (it runs client
code in pcall). If something abnormal happens, like a memory allocation
error, better break the loop and stop the state machine.

Also, raise an exception if we failed to decode an IPROTO_EVENT packet -
this means that something is wrong with the other end and we better stop
the connection (this is what we do if we fail to decode other packets).