test: fix flaky net.box_tx_timeout.test
The problem was that when we wanted to check that
the transaction rolled back after the timeout, we
called `fiber.sleep` on the local instance, while
the timeout for transaction counted on the remote
instance. Fix the text, so now we call `fiber.sleep`
on remote server, ensuring that timeout for transaction
expires.

Closes #6586