Add audit log implementation stub
To build audit log as a part of the box library, set the following cmake
variables:
 - ENABLE_AUDIT_LOG: ON
 - AUDIT_LOG_SOURCES: audit log source files
 - EXTRA_BOX_INCLUDE_DIRS: header files needed for compilation

and add "audit_impl.h" to EXTRA_BOX_INCLUDE_DIRS - then it will be
included by "audit.h".

If ENABLE_AUDIT_LOG is unset, then a stub implementation is built, which
spits a warning to the log on an attempt to configure audit log.