debian: eliminate dependency from binutils package
tarantool debian package MUST NOT depends on binutils package. This is
due to the fact that binutils include linker and assembler, what in most
cases forbidden on production servers.

This dependency is a residual from times, when tarantool did use libbfd
for stack unwinding. Now it simply does not required at all.

Fixes #6699