build: fix build with glibc-2.34
Macros SIGSTKSZ used to be an integral constant but
in glibc-2.34 it turned into a runtime function so it
cannot be used as constant known size for arrays anymore.

Beyond this, SIGSTKSZ is not enough for alt. signal stack size
when you use ASAN, so the size was increased.

Closes #6686

(cherry picked from commit 9c01b325a397df4706e9c2ff4e867bcbee3020c1)